requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
blas=1.0=mkl
bottleneck=1.3.7=py312ha883a20_0
brotli=1.0.9=h5eee18b_8
brotli-bin=1.0.9=h5eee18b_8
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.7.2=h06a4308_0
cairo=1.16.0=hb05425b_5
contourpy=1.2.0=py312hdb19cb5_0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
dbus=1.13.18=hb2f20db_0
eigen=3.4.0=hdb19cb5_0
expat=2.6.2=h6a678d5_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.13.1=py312h06a4308_0
fontconfig=2.14.1=h55d465d_3
fonttools=4.51.0=py312h5eee18b_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.3.1=py312h06a4308_0
future=1.0.0=pypi_0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
graphite2=1.3.14=h295c915_1
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h5py=3.12.1=pypi_0
harfbuzz=4.3.0=hf52aaf7_2
hdf5=1.12.1=h2b7332f_3
icu=73.1=h6a678d5_0
intel-openmp=2023.1.0=hdb19cb5_46306
iso8601=2.1.0=pypi_0
jinja2=3.1.4=py312h06a4308_0
jpeg=9e=h5eee18b_3
keras=3.6.0=pypi_0
kiwisolver=1.4.4=py312h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libabseil=20240116.2=cxx17_h6a678d5_0
libbrotlicommon=1.0.9=h5eee18b_8
libbrotlidec=1.0.9=h5eee18b_8
libbrotlienc=1.0.9=h5eee18b_8
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcups=2.4.2=h2d74bed_1
libcurl=8.9.1=h251f7ec_0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libllvm14=14.0.6=hdb19cb5_3
libnghttp2=1.57.0=h2d74bed_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.39=h5eee18b_0
libpq=12.17=hdbd6064_0
libprotobuf=4.25.3=he621ea3_0
libssh2=1.11.0=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libvpx=1.7.0=h439df22_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h097e994_2
libxml2=2.13.1=hfdd30dd_2
lz4-c=1.9.4=h6a678d5_1
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py312h5eee18b_0
matplotlib=3.8.4=py312h06a4308_0
matplotlib-base=3.8.4=py312h526ad5a_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_1
mkl_fft=1.3.8=py312h5eee18b_0
mkl_random=1.2.4=py312hdb19cb5_0
ml-dtypes=0.5.0=pypi_0
mpmath=1.3.0=py312h06a4308_0
mysql=5.7.24=h721c034_2
namex=0.0.8=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.3=py312h06a4308_0
numexpr=2.8.7=py312hf827012_0
numpy=1.26.4=py312hc5e2394_0
numpy-base=1.26.4=py312h0da6c21_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
opencv=4.10.0=py312h907dce5_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.0.15=h5eee18b_0
optree=0.13.0=pypi_0
packaging=24.1=py312h06a4308_0
pandas=2.2.2=py312h526ad5a_0
pcre2=10.42=hebb0a14_1
pillow=10.4.0=py312h5eee18b_0
pip=24.2=py312h06a4308_0
pixman=0.40.0=h7f8727e_1
ply=3.11=py312h06a4308_1
pygments=2.18.0=pypi_0
pyparsing=3.0.9=py312h06a4308_0
pyqt=5.15.10=py312h6a678d5_0
pyqt5-sip=12.13.0=py312h5eee18b_0
pyserial=3.5=pypi_0
python=3.12.4=h5148396_1
python-dateutil=2.9.0post0=py312h06a4308_2
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2024.1=py312h06a4308_0
pyyaml=6.0.2=pypi_0
qt-main=5.15.2=h53bd1ea_10
readline=8.2=h5eee18b_0
rich=13.9.2=pypi_0
scipy=1.14.1=pypi_0
setuptools=72.1.0=py312h06a4308_0
sip=6.7.12=py312h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h5eee18b_0
sympy=1.13.1=pypi_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.14=h39e8969_0
torch=2.5.0=pypi_0
torchvision=0.20.0=pypi_0
tornado=6.4.1=py312h5eee18b_0
tqdm=4.66.4=py312he106c6f_0
triton=3.1.0=pypi_0
typing_extensions=4.11.0=py312h06a4308_0
tzdata=2024a=h04d1e81_0
unicodedata2=15.1.0=py312h5eee18b_0
wheel=0.43.0=py312h06a4308_0
x264=1!157.20191217=h7b6447c_0
xz=5.4.6=h5eee18b_1
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
